@@ -1,20 +1,33 @@
 {
   "name": "now-client-camera",
-  "version": "0.3.0",
-  "private": true,
-  "scripts": {
-    "eslint": "./node_modules/eslint/bin/eslint.js .",
-    "start": "./node_modules/babel-cli/bin/babel-node.js app.js"
+  "version": "0.4.0",
+  "description": "CAMPHOR- Now Camera Client",
+  "preferGlobal": true,
+  "bin": {
+    "now-client-camera": "./app.js"
   },
+  "scripts": {
+    "build": "./node_modules/babel-cli/bin/babel.js src -d lib",
+    "watch": "./node_modules/babel-cli/bin/babel.js --watch src -d lib",
+    "eslint": "./node_modules/eslint/bin/eslint.js app.js ./src",
+    "start": "./app.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/camphor-/now-client-camera"
+  },
+  "bugs": {
+    "url": "https://github.com/camphor-/now-client-camera/issues"
+  },
+  "license": "MIT",
   "dependencies": {
-    "babel": "~6.5.1",
-    "babel-cli": "~6.5.1",
-    "babel-preset-es2015-node5": "~1.1.2",
     "debug": "~2.2.0",
     "socket.io-client": "~1.4.5",
-    "yargs": "~4.2.0"
+    "yargs": "~4.4.0"
   },
   "devDependencies": {
-    "eslint": "~2.2.0"
+    "babel-cli": "~6.7.5",
+    "babel-preset-es2015": "^6.6.0",
+    "eslint": "~2.7.0"
   }
 }
