@@ -1,16 +1,6 @@
-'use strict';
+import {exec} from 'child_process';
 
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-
-var _child_process = require('child_process');
-
-var _base = require('./base');
-
-var _base2 = _interopRequireDefault(_base);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+import BaseDriver from './base';
 
 const MAX_BUFFER = 1 * 1024 * 1024; // 1MiB
 
@@ -26,14 +16,14 @@ const parseInteger = (string, defaultValue) => {
 const WIDTH = parseInteger(process.env.RASPISTILL_WIDTH, 320);
 const HEIGHT = parseInteger(process.env.RASPISTILL_HEIGHT, 240);
 
-exports.default = class extends _base2.default {
+export default class extends BaseDriver {
   static get name() {
     return 'raspistill';
   }
 
   take() {
     return new Promise((resolve, reject) => {
-      (0, _child_process.exec)(`raspistill -w ${ WIDTH } -h ${ HEIGHT } -n -e jpg -o -`, {
+      exec(`raspistill -w ${WIDTH} -h ${HEIGHT} -n -e jpg -o -`, {
         encoding: 'buffer',
         maxBuffer: MAX_BUFFER
       }, (err, stdout, stderr) => {
@@ -45,4 +35,4 @@ exports.default = class extends _base2.default {
       });
     });
   }
-};
+}
